requirements.txt
Django>=2.2.2, <2.2.3
djangorestframework>=3.10.1,<3.10.2
flake8>=3.6.0,<3.7.0
PyYAML>=5.1.1,<5.1.2
django-rest-swagger